The following updates have been added to the UserDetail page:
Basic information and a list of products are displayed for the user's profile.

The user's profile picture and username are shown.

The average rating received by the user and the number of reviews are displayed on the profile page.

The products sold by the user are listed, and the number of products is shown in the profile title.

The list of products is automatically updated.